rename `iterator` module to `stepper`
